moved to other fields test file
